+++ tests/testthat/test-distributions.R
@@ -29,20 +29,44 @@ test_that("dsem example is working ", {
   fit1 = dsem( 
     tsdata = ts(data.frame(x=z_ti[,1], y=z_ti[,2])),
     sem = sem,
-    family = c("fixed", "fixed") 
+    family = list( x = fixed(), y = fixed() )
   )
   fit2 = dsem( 
     tsdata = ts(data.frame(x=y_ti[,1], y=y_ti[,2])),
     sem = sem,
-    family = c("poisson", "poisson") 
+    family = list( x = poisson("log"), y = poisson("log") )
   )
-  expect_equal( as.numeric(fit2$opt$obj), as.numeric(fit2$opt$obj), tolerance=1e-2 )
+  #expect_equal( as.numeric(fit2$opt$obj), as.numeric(fit2$opt$obj), tolerance=1e-2 )
 
   if( FALSE ){
-    fit3 = dsemRTMB( tsdata = ts(data.frame(x=y_ti[,1], y=y_ti[,2])),
-          sem = sem,
-          family = c("poisson", "poisson") ) 
+    fit3 = dsemRTMB(
+             tsdata = ts(data.frame(x=y_ti[,1], y=y_ti[,2])),
+             sem = sem,
+             family = list( x = poisson(), y = poisson() )
+          )
     expect_equal( resid(fit2), resid(fit3), tolerance=1e-2 )
   }
 })
 
+#
+test_that("dsem `gaussian_fixed_sd` is working ", {
+  set.seed(123)
+  z = rnorm(100)
+  y = 1 + 1 * z + rnorm(100)
+  x = z + rnorm(100)
+
+  fit0 = dsem(
+    sem = "x -> y, 0, beta",
+    tsdata = ts( data.frame(x=x, y=y) ),
+    family = list( x = fixed(), y = fixed() )
+  )
+
+  fit1 = dsem(
+    sem = "x -> y, 0, beta",
+    tsdata = ts( data.frame(x=x, y=y) ),
+    family = list( x = gaussian_fixed_sd(sd=1), y = fixed() )
+  )
+
+  # Should be identical (despite different slopes) because it's using a measurement-error form for identical model
+  expect_equal( as.numeric(fit0$opt$obj), as.numeric(fit1$opt$obj), tolerance=1e-2 )
+})